Updates - Renamed and reorganized files    * GameManager.js changed to MineSweeper.js    * minesweeper.js changed to GameManager.js - Added a way to reload game without page reload - Fixed issue of music not playing without page reload
