+++ games/minesweeper/js/GameManager.js
@@ -1,41 +1,29 @@
-import GameBoard from './GameBoard.js';
+import MineSweeper from './MineSweeper.js';
 import Modal from './Modal.js';
+import MusicPlayer from './MusicPlayer.js';
 
-export default class GameManager {
-  constructor(size, numBombs) {
-    this.flagsRemaining = numBombs;
-    this.size = size;
-    this.numBombs = numBombs;
-    this.flaggedBombs = 0;
-    this.time = { m: 0, s: 0 };
-    this.state = 'playing';
+class GameManager {
+  constructor() {
+    this.musicPlayer = new MusicPlayer();
+    document.getElementById('container').style.display = 'none';
     this.modal = null;
-    this.gb = new GameBoard(
-      'board',
-      20,
-      (pos) => this.click(pos),
-      (pos) => this.rightClick(pos)
-    );
-    this.placeBombs();
-    document.getElementById('num-flags').innerText = this.flagsRemaining;
-    document.getElementById('num-bombs').innerText = this.numBombs;
-    document.getElementById('size').innerText = `${this.size}x${this.size}`;
-    document.getElementById('time').innerText = '00:00';
-    setTimeout(() => this.updateTime(), 1000);
-  }
-
-  updateTime() {
-    if (this.state === 'playing') {
-      this.time.s++;
-      if (this.time.s > 60) {
-        this.time.s = 0;
-        this.time.m++;
+    let tn = document.createTextNode(`Wanna play a game?`);
+    this.modal = new Modal(tn);
+    this.modal.addButton(
+      'New Game',
+      (e) => {
+        this.musicPlayer.addTrack('./media/music/looking_for_adventure.wav');
+        this.musicPlayer.addTrack('./media/music/Epic_Sport.wav');
+        this.musicPlayer.addTrack('./media/music/Inspiring_Muted_Guitar.mp3');
+        this.musicPlayer.play();
+        document.getElementById('container').style.display = 'flex';
+        this.newGame();
+      },
+      {
+        color: 'blue',
+        textColor: 'white',
       }
-      document.getElementById('time').innerText = `${this.time.m
-        .toString()
-        .padStart(2, '0')}:${this.time.s.toString().padStart(2, '0')}`;
-    }
-    setTimeout(() => this.updateTime(), 1000);
+    );
   }
 
   exitClicked() {
@@ -43,13 +31,23 @@ export default class GameManager {
   }
 
   newGame() {
-    window.location.reload();
+    if (this.modal) {
+      this.modal.remove();
+    }
+    if (this.game != undefined) {
+      this.game.removeBoard();
+    }
+    this.game = new MineSweeper(
+      20,
+      5,
+      (time) => this.win(time),
+      () => this.lose()
+    );
   }
 
-  win() {
-    this.state = 'ended';
+  win(time) {
     let tn = document.createTextNode(
-      `Congrats! You won in ${this.time.m} minute(s) and ${this.time.s} second(s)`
+      `Congrats! You won in ${time.m} minute(s) and ${time.s} second(s)`
     );
     var audio = new Audio('./media/sounds/Victory.mp3');
     audio.play();
@@ -64,84 +62,19 @@ export default class GameManager {
     });
   }
 
-  lose(pos) {
-    this.state = 'ended';
-    this.gb.explode(pos);
-    var audio = new Audio('./media/sounds/Explosion Powerful.mp3');
-    audio.play();
-    setTimeout(() => {
-      let gif = document.createElement('img');
-      gif.src = './media/images/lose.gif';
-      this.modal = new Modal(gif);
-      this.modal.addButton('Wallow in defeat!', (e) => this.exitClicked(), {
-        color: 'red',
-        textColor: 'white',
-      });
-      this.modal.addButton('Walk of Shame', (e) => this.newGame(), {
-        color: 'blue',
-        textColor: 'white',
-      });
-    }, 2000);
-  }
-
-  rightClick(pos) {
-    if (this.gb.isFlagged(pos)) {
-      this.gb.toggleFlag(pos);
-      if (this.gb.isBomb(pos)) this.flaggedBombs--;
-      this.flagsRemaining++;
-    } else if (!this.gb.isRevealed(pos)) {
-      this.gb.toggleFlag(pos);
-      if (this.gb.isBomb(pos)) this.flaggedBombs++;
-      this.flagsRemaining--;
-      if (this.flagsRemaining === 0 && this.flaggedBombs === this.numBombs)
-        this.win();
-    }
-    document.getElementById('num-flags').innerText = this.flagsRemaining;
-  }
-
-  click(pos) {
-    //doMove(gb, pos);
-    if (this.state != 'ended') {
-      if (this.gb.isFlagged(pos)) {
-        this.gb.toggleFlag(pos);
-        if (this.gb.isBomb(pos)) this.flaggedBombs--;
-        this.flagsRemaining++;
-      } else if (this.gb.isBomb(pos)) {
-        this.lose(pos);
-      } else {
-        this.doMove(pos);
-      }
-    }
-  }
-
-  placeBombs() {
-    let size = this.gb.size;
-    for (let i = 0; i < this.numBombs; i++) {
-      while (true) {
-        let x = Math.floor(Math.random() * this.size);
-        let y = Math.floor(Math.random() * this.size);
-        if (this.gb.isBomb({ y: y, x: x }) === 1) continue;
-        this.gb.addBomb({ y: y, x: x });
-        break;
-      }
-    }
-  }
-
-  doMove(pos) {
-    if (
-      !this.gb.isInBounds(pos) ||
-      this.gb.isRevealed(pos) ||
-      this.gb.isExploded(pos)
-    )
-      return;
-    this.gb.setState(pos, this.gb.getState(pos) | GameBoard.REVEALED);
-    if (this.gb.countBombs(pos) === 0) {
-      setTimeout(() => {
-        this.doMove({ y: pos.y + 1, x: pos.x });
-        this.doMove({ y: pos.y - 1, x: pos.x });
-        this.doMove({ y: pos.y, x: pos.x + 1 });
-        this.doMove({ y: pos.y, x: pos.x - 1 });
-      }, 100);
-    }
+  lose() {
+    let gif = document.createElement('img');
+    gif.src = './media/images/lose.gif';
+    this.modal = new Modal(gif);
+    this.modal.addButton('Wallow in defeat!', (e) => this.exitClicked(), {
+      color: 'red',
+      textColor: 'white',
+    });
+    this.modal.addButton('Walk of Shame', (e) => this.newGame(), {
+      color: 'blue',
+      textColor: 'white',
+    });
   }
 }
+
+let gm = new GameManager();
