+++ games/minesweeper/minesweeper.html
@@ -5,8 +5,8 @@
     <link rel="stylesheet" href="./styles.css" />
   </head>
   <body>
-    <div class="container">
-      <div id="board" class="game-board"></div>
+    <div id="container" class="container">
+      <div id="board"></div>
       <div class="sidebar">
         <div>
           <label class="info-label">Time</label>
@@ -29,7 +29,7 @@
         </div>
       </div>
     </div>
-    <script type="module" src="./js/minesweeper.js"></script>
+    <script type="module" src="./js/GameManager.js"></script>
     <script language="Javascript"></script>
   </body>
 </html>
