+++ games/minesweeper/styles.css
@@ -25,6 +25,8 @@ body {
 .container {
   display: flex;
   position: absolute;
+  justify-content: center;
+  width: 100%;
 }
 
 .sidebar {
